Player movement, preloader, bullet to bullet collisions, bullet velocity
I changed the player movement so that it applies a lot more thrust but only when the body is under a max velocity.
This has given what I think is overall better movement.
However, when the player reaches a higher velocity than maxVel (like if it's hit by a fast asteroid) it's not allowed to move.
This gives the feeling of being stunned by the hit and I could keep it as a feature. If I decide to not keep it I should add a check between the body.rotation and direction.angle, so if the player rotation is not in the direction we are going we can still apply the force to move.